requirements.txt
django
gunicorn
psycopg2
djangorestframework
django-cors-headers
django-extensions
datetimerange
whitenoise
requests
ipdb